moved user related services to UserService class
